+++ src/m_config.h
@@ -179,7 +179,7 @@ extern char     *r_lowpixelsize;
 extern bool     r_mirroredweapons;
 extern bool     r_percolumnlighting;
 extern bool     r_pickupeffect;
-extern bool     r_playersprites;
+extern bool     r_playerweapon;
 extern bool     r_radsuiteffect;
 extern bool     r_randomstartframes;
 extern bool     r_rockettrails;
@@ -791,7 +791,7 @@ enum
 
 #define r_pickupeffect_default              true
 
-#define r_playersprites_default             true
+#define r_playerweapon_default              true
 
 #define r_radsuiteffect_default             true